Solve 416. Partition Equal Subset Sum
- Recurse in the same way as before but cache results in a
defaultdict once a result is determined so that we can refer to
it and don't have to recalculate it.

- For the caching in the defaultdict, let the following be the
structure: {idx : {target : True/False} }. idx refers to the fact
that we are considering the input list starting from idx. target
refers to an attempt to achieve a sum of 'target' using the input
list starting from idx. The final value True/False denotes whether
or not it is possible to achieve 'target' using the input list
starting from idx.

- Also, if a True is returned at any point in time, echo that back
to the top since that is sufficient and we don't need to do any
further computations.

- Seems like a pretty nice idea. Do a regular recursive solution
and then find repeated computations and cache them. I wonder if
that's the way for all of these dynamic programming questions??