621 - Try out the question from scratch again
- Use a global timestamp to keep track of the overall time. Add
tasks to the heap using this global timestamp. A task is allowed
after n + 1 steps after the current time. We want to use the
tasks with higher count early and as often as possible.

- This avoids having to deal with the issue of needing to track
cooldown times and then especially, of having to decrement them
for ALL the tasks in every iteration of the loop.

- This works better but some wrong answers are still being
returned. The issue seems to be that when it's time for a task
that was added to heap after being worked on to be used again,
it doesn't end up being the top task in the heap if we have
tasks that have never been worked on with a permitted time
value of 0 in the heap. These 0 permitted time tasks end up
taking precedence over tasks for whom we are ready to work
on again.