621 - Use two heaps for the tasks
- Use one heap for keeping track of the tasks that have been
performed by the CPU at least once. In the other 'leftover' heap,
place the tasks that haven't been performed by the CPU at all.

- To start, put the task with the highest count in the CPU heap
since we want to start working on that task at the earliest.

- Use an overall timestamp to keep track of where we are in the
process.

- The general idea is that we want to start the tasks with the
highest counts early and that we want to do them as often as
possible too since they're the ones that are going to take the
longest to finish.

- While iterating over the CPU heap, check if the 'top' task
has an allowed/permitted time that is less than the overall
time. If it doesn't, that means that we need to wait a little
bit more before we can do that task.

- Look to the leftover heap then. Pick the highest count task
from there and work on that. Since the leftover heap contains
tasks that have never been worked on, there are no restrictions
in terms of any waiting time.

- If the CPU's top task is not permitted yet and the leftover
heap doesn't contain anything, then the CPU must remain idle.

- One large testcase that used to timeout before is now passing.

57 / 71 test cases passed.

- Something seems to be going wrong somewhere..