621 - Check for non-emptiness of leftover heap too
- We keep iterating while either of the heaps has tasks in it.

- Refactor the code a bit to simplify things. We use the CPU heap
if the CPU heap is not empty and if the overall time is at least
as much as the top task in the CPU heap. Otherwise, we look towards
using the tasks from the leftover heap. If the leftover heap is
empty however, the CPU then must remain idle.

- Also, once a task is worked upon, we put it back in the CPU heap
if it's count hasn't become 0. And we update the time for it to
be n + 1 steps away from the current overall time, since that's the
amount of time we need to wait before being able to perform this
task again.

- Remove the code for the previous attempts.

63 / 71 test cases passed.

- There is a test case that's failing wherein it is more preferable
to evenly use the tasks rather than focusing on using the higher
priority tasks to the maximum extent. I think the idea with that is
that if the tasks are evenly spread out, they provide a buffer to
each other during the cooldown/waiting period.

- But, this even distribution approach doesn't seem to work for
other test cases.. Probably need to bring these two ideas together
somehow.