347 - Write most of the implementation for the heap from scratch
- Elements are added to the heap as tuples of (num, count) wherein
we use 'count' as the key for the ordering of the elements.

- For adding to the heap, simply add the new tuple to the end of
the heap and then bubble up the new tuple to the appropriate place
in the heap.

- For popping from the heap, pop the element at the beginning of
the heap. Replace it with the last element of the heap. And then
bubble down that new root element to the appropriate place in the
heap.

- Getting a runtime error while bubbling down right now.