Solve 347. Top K Frequent Elements by implementing heaps from scratch*
- Add two missing things: the size should be reset to k once we're
done popping and bubbling down. Otherwise, we can get an index out
of bounds error when bubbling up since it uses the 'size' variable.

- When bubbling down, we might end up reaching the bottom layer of
the heap i.e. at a leaf node. A leaf node wouldn't have left or
right children. Add a check to see if we're at a leaf node. If we
are, then it's not possible to bubble down further and the element
is at the correct place in the heap. Finish bubbling down.