Solve 49. Group Anagrams
- Words will have the same anagrams if they have the same counts
for the same letters. Create a dictionary where the key will be
the Counter for a string and the value will be the list of strings
that yield that Counter.

- Since Counter objects are not hashable, obtain the (key, value)
pairs of (letter, count), sort them so that same Counters with
different orderings are all brought together, and then create a
tuple out of them so that we get a key that's hashable for the
dictionary.