647 - Attempt to solve the problem using binary search
- Start from the entire string. Find the middle index. If the
characters surrounding the middle are equal, update the palindromes
and continue moving outwards until the characters are equal.

- When we get to a point where the characters are not equal, split
into a left and right binary search on either side of the middle
to find palindromes from slices of the string.

- Maintain a visited set so that we don't visit combinations of
(start_index, stop_index) repeatedly.

- This seems to be working for palindromes of odd length but is
failing for those of even length.

15 / 130 test cases passed.