40 - Mostly solve the problem using approach from another problem
- Solve the problem for the most part using the approach used for
416. Partition Equal Subset Sum.

- Do things in the same way as in starting from the beginning of
the array and building up answers to whether it is possible to
reach a target t using the array starting from idx - so, pairs of
(idx, t) in the defaultdict.

- Pass an additional 'path' variable that will be used to keep
track of the current path along our DFS traversal. If the path
ever reaches a point where the target is 0, append the path to
the list of all paths.

- Add some additional logic to the stuff from 416 since we want
to know _all_ the paths, not just whether a path exists or not.
For example, in the two branches where we're recursing into
(idx + 1, t), once while including the current number and once
while excluding it, add an additional condition to the if block.
If the target is not in the defaultdict, then we need to recurse
into idx + 1, same as before. But also now, if the sub-answer
from idx + 1 is True, we want to recurse into it so that we can
obtain the valid path.

- Sort the candidates before starting because there was an
observation that same paths were being added to the list (these
same paths just had different orders for the same numbers). By
sorting the numbers, I observed that the same paths now had the
same orders as well (It isn't too bad to sort the candidates at
all since #candidates is upper bounded by 100 in the problem).

- In conjuction with the sorting, let the list of all paths be
a set so that only unique paths are stored for the final result.

- Also, swap the positions of the base cases. There might be cases
where a target of 0 is achieved when we reach the end of candidates.
In that case, we would like to add the path obtained in this
situation, but 'candidates[idx:]' would cause the if-check to fail.
Therefore, check for target == 0 first.

172 / 175 test cases passed.

- Getting Time Limit Exceeded. I suspect this is happening in the
recursion step wherein we are redeveloping paths for
(idx + 1, target) when we already know it to be True from a previous
recursive call into it.