40 - Refactor and stop the recursive partitioning earlier
- I sorted the candidates at the beginning. Instead of stopping
the recursion when the target goes below 0, we can stop earlier.

- If the target is less than the candidate at the current index
idx, then due to the candidates being sorted at the beginning, the
target will definitely be less than the candidates that lie ahead.
So, there would be no point in checking further for this specific
target and we can stop partitioning.

- Switch the position of the two clauses in the if condition.
We are checking if the target is less than the candidate at index
idx. idx might be out of bounds for the list as we recurse further
and further. Therefore, put the check for the non-emptiness of the
slice of the list first. Now, we will only compare the target and
the current candidate if that check passes and thereby avoid out
of bounds errors.

- This improves over the previous running time of ~120 ms, 18% to
82 ms, 57%.