Solve 216. Combination Sum III
- Solve this by extending the work done for 40. Combination Sum
II. We are now interested in a 3rd variable besides idx and
target - the path size. We want all k-length paths that sum to
the given target n.

- Extend the defaultdict to have another layer of depth for the
path size. So, the structure will be:
{idx: target: path_size: [] }}}, which means that, considering
the input list starting from idx, we are searching for target
and that the path so far has length path_size. The value which
is a list will contain all the paths that are valid under these
constraints.

- It builds on 40. Combination Sum II since we now have an extra
restriction about the size of the path. 40 didn't have any such
constraints about the size.

- In the base case, if we change it to path_size <= size_limit,
we would get paths of all lengths less than the size limit
that allow us to get to the target.

- The container doesn't need to be a set since we don't have
duplicate numbers within our candidates. A list is sufficient.

- The desired base case is when the target reaches 0 and the
path length is equal to k.

- Add an additional check for if the path length goes beyond k.

- The sub-results are incorporated into the 'main' result in
the same way as before. We simply have 3 values to depend on
now instead of 2 when fetching the sub-results.

- Solved this pretty quickly too! 17 minutes after the initial
commit. The running time is super nice too! 28 ms, 91% !!