216 - Refactor path_size to start from k instead of 0
- Change the way path_size is being treated. Use it in the same
way like target is being used - as a resource. Every time we
include the number at idx and we recurse, there is one less spot
available in the path and therefore we decrement the path size.

- So, start from the full availability for path_size, which is k.
The desired base case is when we're able to reach a target of 0
and all the spots in the path have been used up.

- For the other base case, prevent recursing further if the path
size becomes 0 i.e. there are no more spots available to
construct the current path.

- When the number at idx is included, use a decremented version
of path_size for recursing and finding the sub-results.

- This idea puts target and path size in sync. They both start
out 'full' and then they dwindle down until we get the desired
result of 0, 0 for them.