377 - Find the #permutations of each of the paths generated
- Using the results of 40. Combination Sum II, we obtain all the
unique paths that lead to the target such that each number is only
used once in each path.

- Then, we find the number of permutations that are possible for
each of the paths, taking into account any duplicate numbers in
the path.

- This is actually insufficient for this problem since the problem
allows for a number to be used more than once in a path. Need to
revise this/use something else.

- Also, there was no need to handle duplicates in the path. The
input contains distinct numbers so it's not possible for a path
to contain duplicate numbers.