Solve 377. Combination Sum IV
- The time limit was being exceeded in the last commit since it
was quite time consuming to build up the paths as we moved towards
higher and higher targets.

- Since we're only interested in the #permutations of all the
paths, only store the #paths in the list as opposed to the paths
themselves.

- So, for the base case when tgt - candt == 0, store 1 instead of
the lone-path candidate. And when we're using a cached result in
the other if-block, simply update the #paths using the sub-result
as opposed to appending new paths.

- Since this is bottom-up dynamic programming, sort the numbers.
We need to do that because a candidate will use results from
candidates that are smaller than it.

- Add another if-block which checks if the candidate is larger
than the target. If so, then there is no point in checking for
that specific target anymore since the candidates that come after
are larger anyways due to the sorting. So, break out of the loop
for that target at this point.

- Running time: 44 ms, 51%