785 - Write a regular BFS traversal for a disconnected graph
- Start a BFS from a random node, mark it to belong to one set
(use values of 0, 1 for the two sets), flip the set value so
that the next nodes will belong in the opposite set, add the
neighbors to the BFS queue. If a neighbor belongs to a set
already and it is the same set as the current node, the graph
is not bipartite and we return False.

- For the first example test case, this returns False as expected
but it also returns False for the second example test case when
it should be True instead. The set values don't seem to be
getting set properly for a node's neighbors.