785 - Change up when the sets are marked for the nodes
- The current location in the traversal where the set is being
marked isn't working. Move it to different place(s).

- For the start node of a BFS, mark it's set right at the
beginning before starting the BFS.

- Add a new condition for the neighbors. We only want to add them
to the BFS queue if they haven't been added before so that we only
traverse over a node once.

- Once that condition passes, remove the neighbor from the list
of all nodes, add it to the queue and mark the set for it, which
will be the inverse of the set for the current node.

- After the loop for the neighbors of the current node is finished,
invert the set value. We do this _after_ the loop so that all the
neighbors get the same set value w.r.t the current node.

65 / 80 test cases passed.

- Wrong answer for test case: [[1,4],[0,2],[1],[4],[0,3]].
Expected answer is True.