Solve 886. Possible Bipartition
- Create an adjacency list graph representation from the edge list
'dislikes' and use the solution from 785. Is Graph Bipartite?

- For the dislikes, we just want to check if they can be separated
into two groups or not and that anybody who dislikes anybody else
must have a different group number, which is the same idea as
assigning different groups to connected nodes to try and create a
partition to get a bipartite graph.

- Runtime: 700 ms, 74%