684 - Partially solve the problem
- The idea I'm using here is that any edge that doesn't contribute
a node that hasn't been seen before is the extra edge. There is
only extra edge. That means that all the other edges are being
used to add at least one new node to the tree.

- If there is an edge both of whose endpoints are already in the
set of visited nodes, that edge will be the extra edge. And, since
we're looking for the last such edge in the input in case of
multiple answers, the for-loop suffices since we would update the
extra edge if a new valid edge was found down the line.

21 / 39 test cases passed.

- Failing for test case:
[[9,10],[5,8],[2,6],[1,5],[3,8],[4,9],[8,10],[4,10],[6,8],[7,9]]
Output: [6, 8]
Expected: [4, 10]