684 - Attempt 2 by checking the impact of removing each edge
- First, construct the adjacency list representation for the
undirected graph. Then, for each input edge, check if the nodes
in question still have neighbors left after we remove the ends
of that edge.

- If they still have neighbors, then this edge is not contributing
anything and it can be considered as the extra edge.

- Since we want the last extra edge, a simple for-loop suffices
wherein we update the extra edge when the condition is satisfied.

21 / 39 test cases passed.

- Failing for the same test case as before:
[[9,10],[5,8],[2,6],[1,5],[3,8],[4,9],[8,10],[4,10],[6,8],[7,9]]
Output: [6,8], Expected: [4,10].

- This solution selects an extra edge based on the fact that are
there neighbors still left after the removal of that edge. But,
that is not enough because the impact of removing that edge could
be that you get two disconnected parts of the whole graph. So,
those two nodes will still have neighbors but the graph would
then have been broken up into 2 parts.

- An alternative I could try is to run a BFS and assign some
kinds of numbers to the nodes.