Solve 684. Redundant Connection
- Solve the problem by building components within the graph.
Maintain data in 2 formats: a defaultdict components that will
store all the nodes that belong to a component numbered x and
another defaultdict which will contain the component number for
each node (starting out from a default of positive infinity).

- This will establish a two-way view of the data wherein we
can e.g. get all the nodes that belong to component #2 as well
as component numbers (CNs) for individual nodes.

- The goal here is to iterate over the edges and assign
CNs to the nodes. We want all the nodes to end up in CN 1
in the end. Any time an edge connects two nodes from
different components, the nodes within both components
are now reachable from one another and we assign all of them
the smaller CN between the two components.

- While this process is happening, we are on the lookout for
any edge that has nodes who have the same CNs. Such an edge
is connecting two nodes that are already reachable to one
another and therefore this edge will be the extra edge. We
continue to be on the lookout for such an edge throughout the
whole process since we want the _last_ extra edge.

- As we iterate over the edges and update CNs for the nodes,
also update the nodes that belong to a particular CN (the
two-way view of the data) and clear out a component once the
nodes within it have been assigned a new CN.

- Use some if-else conditions to determine for which nodes we
need to update the CNs as well as which component to clear.

- Runtime: 60 ms, 52%