275 - Partially solve the problem
- The citations are sorted in order. Run a binary search on it.
At the arrived index, check if the #citations that are left
after and including this index are at least as much as the

- For example, if the value at an index is 5, then we want to
check if starting from that index onwards, we can find at least
5 papers or not.

- If we can, then we set the lower bound of the binary search to
this index since there might be values that lie ahead that also
satisfy this condition and have a higher count for #citations.

- If we cannot, then we set the upper bound of the binary
search to this index and search in the left part of the list.

48 / 83 test cases passed.

- Failing for testcase: [0], Output: -1, Expected: 0