275 - Use #papers and #citations together
- Change the idea from the last commit of how the maximum
h-index is being determined. Consider a case where the input
list has 10 papers in it and the very first paper has 11
citations (and since the list is sorted, all the subsequent
papers have >= 11 citations).

- In such a case, the h-index would be 10 since there are 10
papers (the full list) wherein each paper has at least 10
citations. This is actually the base case:
num_citns <= citations[0].

- In the generalized case where all papers may not be usable,
conduct a binary search and check if #papers that are left after
and including this index are less than equal to the #citations
for the paper at that index. This case actually translates to the
base case that we saw above. So, basically, we're performing
binary search to locate the best-performing base case in the
input.

- If the condition is not satisfied, then we've gone too far
left and need to search to the right of that index.

- The stuff from the previous commit was kind of on the right
track but things were in the reverse a bit.

81 / 83 test cases passed.

- Failing for test case: [0,0]. Output: -1, Expected: 0