210 - Use a path to keep track of a valid order for the courses
- We append a course to the path once it's going into status 2
(completed). In a DFS, the deeper nodes will reach completed
status first. We want this here because we want the prerequisites
to be in the path before we put the courses that require those
prerequisites, so that we can get a valid order for the courses.

- 36 / 44 test cases passed.

- Failing test case: 3, [[1, 0]], Output: [0, 1],
Expected: [2, 0, 1]. It seems like there's a problem when there
are nodes/courses that do not have any prerequisites.