997 - Mostly solve the problem
- I approached the problem as a star graph wherein everybody only
trusts the town judge (and the town judge trusts no one). Then,
I search for the number that doesn't have any outgoing connections
by taking a set difference. If there is such a number, that will
be the town judge, else the answer will be -1.

87 / 92 test cases passed.

- Failing for test case: 3, [[1,2],[2,3]]. Output: 3, Expected: -1
3 trusts no one and could be the town judge but _everybody_ must
trust the town judge but here, 1 does not trust 3.